requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
drf_spectacular~=0.26.1
